// src/api/orval/model/issuedOvpnFileDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * DataGate Monitor API
 * OpenAPI spec version: v1
 */

export interface IssuedOvpnFileDto {
  id?: number;
  vpnServerId?: number;
  /** @nullable */
  externalId?: string | null;
  /** @nullable */
  commonName?: string | null;
  /** @nullable */
  certId?: string | null;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  filePath?: string | null;
  issuedAt?: string;
  /** @nullable */
  issuedTo?: string | null;
  /** @nullable */
  pemFilePath?: string | null;
  /** @nullable */
  certFilePath?: string | null;
  /** @nullable */
  keyFilePath?: string | null;
  /** @nullable */
  reqFilePath?: string | null;
  isRevoked?: boolean;
  /** @nullable */
  message?: string | null;
  createDate?: string;
  lastUpdate?: string;
}