// src/api/orval/model/revokeFileRequest.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface RevokeFileRequest {
  /**
   * @minimum 1
   * @maximum 2147483647
   */
  vpnServerId: number;
  ovpnFileId: number;
  /** @minLength 1 */
  commonName: string;
  isRevoked?: boolean;
}